+++ src/rust/ensogl/src/display/shape/primitive.rs
@@ -6,3 +6,4 @@ pub mod system;
 
 pub use def::*;
 pub use shader::*;
+pub use system::*;
