+++ src/rust/ensogl/src/display/shape/primitive/system.rs
@@ -4,10 +4,12 @@
 use crate::prelude::*;
 
 use super::def::*;
+use super::def;
 
 use crate::display;
 use crate::display::shape::primitive::shader;
 use crate::display::symbol::geometry::SpriteSystem;
+use crate::display::symbol::geometry::Sprite;
 use crate::display::symbol::material;
 use crate::display::symbol::material::Material;
 use crate::display::world::World;
@@ -40,7 +42,7 @@ impl CloneRef for ShapeSystemDefinition {
 
 impl ShapeSystemDefinition {
     /// Constructor.
-    pub fn new<'t,S,Sh:Shape>(scene:S, shape:&Sh) -> Self
+    pub fn new<'t,S,Sh:def::Shape>(scene:S, shape:&Sh) -> Self
     where S : Into<&'t Scene> {
         let sprite_system = SpriteSystem::new(scene);
         let material      = Rc::new(RefCell::new(Self::surface_material()));
@@ -62,7 +64,7 @@ impl ShapeSystemDefinition {
     }
 
     /// Replaces the shape definition.
-    pub fn set_shape<S:Shape>(&self, shape:&S) {
+    pub fn set_shape<S:def::Shape>(&self, shape:&S) {
         let code = shader::builder::Builder::run(shape);
         self.material.borrow_mut().set_code(code);
         self.reload_material();
@@ -85,4 +87,83 @@ impl<'t> From<&'t ShapeSystemDefinition> for &'t display::object::Node {
     fn from(shape_system:&'t ShapeSystemDefinition) -> Self {
         shape_system.sprite_system.display_object()
     }
+}
+
+
+
+#[derive(Clone,CloneRef,Debug,Shrinkwrap)]
+#[clone_ref(bound="Params:CloneRef")]
+pub struct ShapeWrapper<Params> {
+    #[shrinkwrap(main_field)]
+    pub params : Params,
+    pub sprite : Sprite,
+}
+
+
+pub trait ShapeSystem {
+    type ShapeDefinition;
+    fn new(scene:&Scene) -> Self;
+    fn new_instance(&self) -> Shape<Self>;
+}
+
+pub type ShapeDefinition<T> = <T as ShapeSystem>::ShapeDefinition;
+
+pub type Shape<T> = ShapeWrapper<ShapeDefinition<T>>;
+
+
+
+#[macro_export]
+macro_rules! shape {
+    (
+        ($($gpu_param : ident : $gpu_param_type : ty),* $(,)?)
+        {$($body:tt)*}
+    ) => {
+
+        // =============
+        // === Shape ===
+        // =============
+
+        #[derive(Clone,Debug)]
+        pub struct ThisShapeDefinition {
+            $(pub $gpu_param : Attribute<$gpu_param_type>),*
+        }
+
+
+        // ==============
+        // === System ===
+        // ==============
+
+        #[derive(Clone,CloneRef,Debug)]
+        pub struct System {
+            pub shape_system : $crate::display::shape::ShapeSystemDefinition,
+            $(pub $gpu_param : Buffer<$gpu_param_type>),*
+        }
+
+        impl ShapeSystem for System {
+            type ShapeDefinition = ThisShapeDefinition;
+
+            fn new(scene:&Scene) -> Self {
+                let shape_system = $crate::display::shape::ShapeSystemDefinition::new(scene,&Self::shape_def());
+                $(let $gpu_param = shape_system.add_input(stringify!($gpu_param),default::<$gpu_param_type>());)*
+                Self {shape_system,$($gpu_param),*}
+            }
+
+            fn new_instance(&self) -> $crate::display::shape::primitive::system::Shape<Self> {
+                let sprite = self.shape_system.new_instance();
+                let id     = sprite.instance_id;
+                $(let $gpu_param = self.$gpu_param.at(id);)*
+                let params = ThisShapeDefinition {$($gpu_param),*};
+                ShapeWrapper {sprite,params}
+
+            }
+        }
+
+        impl System {
+            pub fn shape_def() -> AnyShape {
+                $($body)*
+            }
+
+
+        }
+    };
 }
