+++ src/rust/lib/debug-scenes/src/shapes.rs
@@ -32,7 +32,7 @@ use ensogl_system_web::StyleSetter;
 use ensogl::display::layout::alignment;
 use wasm_bindgen::JsCast;
 use ensogl::display::scene;
-use ensogl::display::scene::{Scene, Component, ShapeSystem, ShapeWrapper, MouseTarget};
+use ensogl::display::scene::{Scene, Component, MouseTarget};
 
 
 #[derive(Clone,Debug)]
